// src/app/shared/model/Basket/nba-game.ts
export class NbaGame {
  GameID: number;
  AwayTeam: string;
  AwayTeamID: number;
  AwayTeamMoneyLine: number;
  AwayTeamScore: number;
  DateTime: Date;
  Day: Date;
  HomeTeam: string;
  HomeTeamID: number;
  HomeTeamMoneyLine: number;
  HomeTeamScore: number;
  IsBettingClosed: boolean;
  LastUpdate: Date;
  PointSpreadAwayTeamMoneyLine: number;
  PointSpreadHomeTeamMoneyLine: number;
  Season: number;
  SeasonType: number;
  StadiumID: number;
  Status: string;
}